Revert "Faster list load"
