+++ front-end/src/StorageTimeSearch/StorageTimeSearchListGroups.js
@@ -2,13 +2,27 @@ import React, { useEffect } from "react";
 import NavBar from "../NavBar";
 import StorageTimeSearchList from "./StorageTimeSearchList";
 import StorageItem from "./StorageItem";
+import axios from "axios";
+import { useState } from "react";
 
 // uses a loop to go through all the items, and returns the general structure for list
 // (with navbar, header, StorageTimeSearchList, and the itemlist)
 
 //FRUITS (CATEGORY NUMBER 0)
-const StorageTimeSearchListFruits = (props) => {
-  const itemList = props.data.map((item) => {
+const StorageTimeSearchListFruits = () => {
+  const [items, setItems] = useState([]);
+
+  useEffect(() => {
+    getData();
+  }, []);
+
+  const getData = async () => {
+    const axiosResult = await axios.get("/storagetimeitems");
+    let data = await axiosResult.data;
+    setItems(data);
+  };
+
+  const itemList = items.map((item) => {
     if(item.category === 0){
       return<StorageItem key={item.id} item={item}></StorageItem>
     }
@@ -25,8 +39,19 @@ const StorageTimeSearchListFruits = (props) => {
 };
 
 //DAIRY (CATEGORY NUMBER 1)
-const StorageTimeSearchListDairy = (props) => {
-  const itemList = props.data.map((item) => {
+const StorageTimeSearchListDairy = () => {
+  const [items, setItems] = useState([]);
+
+  useEffect(() => {
+    getData();
+  }, []);
+
+  const getData = async () => {
+    const axiosResult = await axios.get("/storagetimeitems");
+    let data = await axiosResult.data;
+    setItems(data);
+  };
+  const itemList = items.map((item) => {
     if(item.category === 1){
       return<StorageItem key={item.id} item={item}></StorageItem>
     }
@@ -43,8 +68,19 @@ const StorageTimeSearchListDairy = (props) => {
 };
 
 //GRAIN (CATEGORY NUMBER 2)
-const StorageTimeSearchListGrain = (props) => {
-  const itemList = props.data.map((item) => {
+const StorageTimeSearchListGrain = () => {
+  const [items, setItems] = useState([]);
+
+  useEffect(() => {
+    getData();
+  }, []);
+
+  const getData = async () => {
+    const axiosResult = await axios.get("/storagetimeitems");
+    let data = await axiosResult.data;
+    setItems(data);
+  };
+  const itemList = items.map((item) => {
     if(item.category === 2){
       return<StorageItem key={item.id} item={item}></StorageItem>
     }
@@ -61,8 +97,19 @@ const StorageTimeSearchListGrain = (props) => {
 };
 
 //MEATS (CATEGORY NUMBER 3)
-const StorageTimeSearchListMeats = (props) => {
-  const itemList = props.data.map((item) => {
+const StorageTimeSearchListMeats = () => {
+  const [items, setItems] = useState([]);
+
+  useEffect(() => {
+    getData();
+  }, []);
+
+  const getData = async () => {
+    const axiosResult = await axios.get("/storagetimeitems");
+    let data = await axiosResult.data;
+    setItems(data);
+  };
+  const itemList = items.map((item) => {
     if(item.category === 3){
       return<StorageItem key={item.id} item={item}></StorageItem>
     }
